requirements.txt
streamlit
google-generativeai
langchain
PyPDF2
python-docx
faiss-cpu
langchain_google_genai